// src/models/claim/claim.typings.d.ts
export interface ClaimType {
        id?: Number,
        policyNumber: string, 
        claimNumber: string, 
        pvNumber: string,
        lossLocation: string,
        lossHour: string,
        lossDate: Date,
        authorityName:string,
        responsibilityShare:string,
        accidentDescription:string,
        judicialProcedureId:Number|null,
        createdBy:string,
        updatedBy:string,
        status:string,
        decidedBy:string,
        decidedAt:Date,
        driverId:Number | null,
        authorityType: Number,
        vehiculeId:Number | null,
        rejectionReason: string,
        updatedAt?:Date,
        createdAt?:Date
}